# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movie-reviews-api",
        "description": "API documentation movie-reviews-api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/medias/{media_url}/reviews": {
            "get": {
                "tags": [
                    "review"
                ],
                "summary": "get reviews of specefied media.",
                "operationId": "c1d8245b2e44b34e1fd86821d2800322",
                "parameters": [
                    {
                        "name": "media_url",
                        "in": "path",
                        "description": "url of media",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "search_term",
                        "in": "query",
                        "description": "search_term",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "username",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "filter score",
                        "schema": {
                            "format": "float",
                            "default": ""
                        }
                    },
                    {
                        "name": "likes_count",
                        "in": "query",
                        "description": "filter likes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "dislikes_count",
                        "in": "query",
                        "description": "filter dislikes_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "replies_count",
                        "in": "query",
                        "description": "filter replies_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "number of items in a page",
                        "schema": {
                            "format": "int64",
                            "default": 10
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort by",
                        "schema": {
                            "format": "string",
                            "default": "",
                            "enum": [
                                "newest",
                                "oldest",
                                "highest-score",
                                "lowest-score",
                                "most-likes",
                                "least-likes",
                                "most-dislikes",
                                "least-dislikes",
                                "most-replies",
                                "least-replies"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "review"
                ],
                "summary": "create a review.",
                "operationId": "def228f6316270b177058d0eec163a37",
                "parameters": [
                    {
                        "name": "media_url",
                        "in": "path",
                        "description": "url of media",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    }
                ],
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "review": {
                                        "description": "review text",
                                        "default": ""
                                    },
                                    "score": {
                                        "description": "integer between 0,100",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "review created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "media is already reviewd ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/reviews/{id}": {
            "get": {
                "tags": [
                    "review"
                ],
                "summary": "get specified review.",
                "operationId": "a9949d3a965ed93ae30547c07eef7505",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of review",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "review not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "review found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/{review_id}": {
            "put": {
                "tags": [
                    "review"
                ],
                "summary": "update specified review.",
                "operationId": "3ce2173f9fd4ec1ace584902c213f48b",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "description": "id of review",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "review": {
                                        "description": "new review text",
                                        "default": ""
                                    },
                                    "score": {
                                        "description": "new score: integer between 0,100",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "review updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "review not found or not created by user ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "review"
                ],
                "summary": "delete specified review.",
                "operationId": "8ee3c7225c010f99d60267d72118defe",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "description": "id of review",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "review deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "review not found or not created by user ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/reviews/{review_id}/like": {
            "post": {
                "tags": [
                    "review"
                ],
                "summary": "like/dislike specified review.",
                "operationId": "3113e2af75bf70c2643eb3055112fe72",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "description": "id of review",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "is_liked": {
                                        "description": "",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "review liked/disliked",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "review already liked/disliked",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "review"
                ],
                "summary": "remove like/dislike from specified review.",
                "operationId": "00f017f89c0b607eb9a2b6da2ee0b732",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "description": "id of review",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "like/dislike deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "review is not liked/disliked",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}